requirements.txt
# SignalTrust AI Market Scanner Dependencies
# Python 3.7+

# Web Framework
flask>=2.3.0
flask-cors>=4.0.0

# Data Analysis & AI
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.0

# AI Providers (Optional - install only what you need)
# For OpenAI GPT models
openai>=1.0.0

# For Anthropic Claude models
anthropic>=0.18.0

# Visualization
plotly>=5.14.0

# API & Data Fetching
requests>=2.28.0
beautifulsoup4>=4.11.0

# Cloud Storage
boto3>=1.28.0
python-dotenv>=1.0.0

# Utilities
python-dateutil>=2.8.0
schedule>=1.1.0

# Production Server
gunicorn>=21.2.0

# ML & Meta-Model (Optional - for advanced features)
# xgboost>=2.0.0
# lightgbm>=4.1.0
# shap>=0.43.0
# mlflow>=2.9.0

# RAG & Embeddings (Optional)
# sentence-transformers>=2.2.0
# faiss-cpu>=1.7.4

# Async & High Performance (Optional)
# httpx>=0.25.0
# aiohttp>=3.9.0